requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
matplotlib>=3.8.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Optional for enhanced features
pyyaml>=6.0.0
